poprawka...

+++ instance.h
@@ -13,7 +13,7 @@ class Instance
 private:
   std::string file_name;
   int K, Q;
-  float time;
+  double time;
   vector<bool> served;
   
   class Order
@@ -31,7 +31,7 @@ private:
     int get_due_date();
     int get_service_duration();
     
-    float distance_to(int, vector<Order*>);  		//return distance to next_customer
+    double distance_to(int, vector<Order*>);  		//return distance to next_customer
   };
   
   vector<Order*> orders; 				//list of orders
@@ -39,7 +39,7 @@ private:
   bool is_ok();						//return true if instance can be solved 
   int nearest(int start_cust, int vehicle_capacity);	//return nearest customer number
   bool all_served();					//return true if all customers are served
-  float  itinerary(vector<int> &route);			//found one itinerary
+  double  itinerary(vector<int> &route);			//found one itinerary
   
 public:
   Instance(char* file_name, int K); 			//import data from file
